Fix: the names of libraries have been changed

+++ Assignment2/Assignment2b/header.h
@@ -6,7 +6,7 @@
 
 namespace Assignment2b
 {
-    int rpn(const std::string& input);
+    double rpn(const std::string& input);
 }
 
 #endif // HEADER_H
